+++ metadata.rb
@@ -8,20 +8,16 @@ issues_url       'https://github.com/sous-chefs/nagios/issues'
 source_url       'https://github.com/sous-chefs/nagios'
 chef_version     '>= 14'
 
-depends 'apache2', '>= 4.0'
-depends 'nginx', '>= 7.0'
-depends 'php-fpm', '>= 0.7.9'
+depends 'apache2',  '~> 5.0' # 6.0+ removes all recipes and attributes
+depends 'nginx',    '~> 9.0' # 10.0+ removes all recipes and attributes
+depends 'nrpe'
+depends 'php', '>= 4.0.0'
+depends 'yum-epel'
 depends 'zap', '>= 0.6.0'
-
-%w(php yum-epel nrpe).each do |cb|
-  depends cb
-end
 
 %w(centos oracle redhat).each do |os|
   supports os, '>= 7.0'
 end
 
-supports 'amazon'
-supports 'debian', '>= 8.0'
-supports 'scientific'
-supports 'ubuntu', '>= 14.04'
+supports 'debian', '>= 10.0'
+supports 'ubuntu', '>= 16.04'