+++ recipes/server_package.rb
@@ -30,13 +30,18 @@ when 'debian'
   %w(adminpassword adminpassword-repeat).each do |setting|
     execute "debconf-set-selections::#{node['nagios']['server']['vname']}-cgi::#{node['nagios']['server']['vname']}/#{setting}" do
       command "echo #{node['nagios']['server']['vname']}-cgi #{node['nagios']['server']['vname']}/#{setting} password #{random_initial_password} | debconf-set-selections"
+      sensitive true
       not_if "dpkg -l #{node['nagios']['server']['vname']}"
     end
   end
 end
 
-node['nagios']['server']['packages'].each do |pkg|
-  package pkg
+package node['nagios']['server']['packages']
+
+# File typically exists on Debian
+file "#{node['apache']['dir']}/conf-enabled/#{node['nagios']['server']['vname']}-cgi.conf" do
+  manage_symlink_source true
+  action :delete
 end
 
 directory node['nagios']['config_dir'] do